content: add getting-started page
Add a stub for "Getting Started", an introduction to the bot and how to
add it.

As of now, it just explains how to add the bot and to consider changing
the prefix, as well as enabling all commands.

Side Note:
Due to reasons I cannot explain we keep images used in the respective
articles *next* to them. That way (at least as far as I can tell), they
both work during local development and production.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>