legal: add CC-BY-SA license
After a short internal discussion and no other suggestions on the public
side, we'll be licensing this documentation text under the terms of
CC-BY-SA 4.0.

In lieu of this change, update the README and add a footer to the menu
sidebar to clearly state that the text is licensed under these terms.
While we're at it, also include our Terms of Use and our Privacy Policy,
such that we hopefully have most of the legal headaches out of the way.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>