workflows: rename deploy.yml to hugo.yml
Rename the workflow from deploy.yml to hugo.yml to be more descriptive
of what it is actually doing, given that it only really deploys when a
commit is pushed to the master (i.e. production) branch.

While we're at it, also fix a small typo messing up the HUGO_BASE_URL,
it should be botlabs-gg, not botlabs.gg.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>