meta: add contributing guidelines
Add contributing guidelines, outlining where to go for general support
and bug reports, as well as how to submit a pull request.

We define a specification for commit messages, such that it is easier to
read and parse the git shortlog and thus generate proper changelogs.
This specification is intended to present a basic "skill-check" to avoid
rushed contributions with no real value, but they are by no means meant
to be difficult to work with.

To make sure as many potential contributors as possible find these
guidelines, do link to it from the README, although GitHub should
already prompt contributors to read them, given it is located in any of
the documented locations where GitHub detects this special file.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>