docs/core: document core section (#13)
* docs/main: adjust weight for future chapters

Adjust the page weight to place nicely with future chapters coming after
in the menu hierarchy, otherwise the order would be weird.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* docs/main: fix image URLs in markdown documents

Testing indicated that to properly link to images in markdown documents
intended for content, URLs to that image should start with a leading
"/".

This is especially true for deeply nested pages.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* docs/main: recommend reading the entire docs

Recommend reading the documentation in its entirety before and/or during
setup, to get things started easier.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* docs/core: initial work on core documentation

Initialise the work on documenting the Core section on the control
panel. This initial page succintly summarises what's going on in this
chapter and lists the subpages via the `children` shortcode.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* docs/core: document control panel access

Document the control panel access settings, going through the individual
knobs to twist from top to bottom as seen on the web.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* docs/core: document control panel logs

Document the control panel logs subpage. This page is quite short, as
per the nature of the page itself. There isn't much to say about it, and
its content is self-explanatory nonetheless.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* doc/core: document command settings page

Document the command settings page.

We're detailing the command override priority, how the prefix setting
applies to flags and switches. For available options we first go over
the common options that all three tabs have in common, and then get more
specific via options available in channel overrides as well as those
available in command overrides.

This pretty much concludes the command settings page as well.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* doc/core: squash typos in access settings page

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* fixup! doc/core: document command settings page

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* fixup! doc/core: document command settings page

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>