hugo: remove some useless/ugly themes
Remove most of the themes that have no use beyond changing a single
colour around, or are just not very nice to look at / redundant compared
to other themes.

Internal voting has concluded that we want to keep the Zen theme. A
custom theme based on YAGPDB's colours is also being considered and
should hopefully be the default in the future.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>