notifications: copy current notif docs
Copy the currently notifications docs; internal discussion concluded
that this is the current milestone, such that we have a faster migration
to the statically built site.

Some modifications such as notice shortcodes were necessary for it to
work at all. I also took the liberty and removed links to dead /
redundant / not yet existing links.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>