hugo: rename to help center
In light of the merging of learn into this repository, which also means
both docs and learn will be hosted under the same subdomain, rename the
site to YAGPDB Help Center instead of YAGPDB Docs.

The name is quite obviously up for debate, this is just to get a
starting ground for discussion.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>