learn: migrate introduction to doks
Migrate the learning resources's introduction to Doks.

I took the liberty of merging the first two *very short* pages together,
thus forming a more coherent flow of reading.

Furthermore, I removed the (repeated) introduction to the custom command
interface — we already have those in the documentation, there is
basically no need to have *essentially the same thing* twice.

SUMMARY.md has been deleted as well, as that is just a table of
contents, which we already have on the left side of the site.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>