docs: bring YouTube page up to date
Update the YouTube documentation page, as it is seriously outdated and
thus inaccurate.

List the supported URLs YAGPDB attempts to extract the channel from;
add a separate header for the configuration options and explain how to
edit an already existing feed.

Detail the limit of 10 (250) active feeds and the behaviour when
exceeding that limit.

Explain the custom announcement message, listing additional context data
available therein, as well as outline the "danger" that arises regarding
role / everyone mentions when using it.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>