meta: refine contribution guide
The contributing guide was partly outdated, mentioning to run Hugo for
building / previewing the site and was generally confusingly written,
exhibiting a non-linear style.

Update it such that it can now be read linearly, outlining the new
preview / build process, explaining the markdown features we want
contributors to use. We also now have a quick rundown of the superficial
repository structure, hopefully enabling contributors to find relevant
files faster when they're viewing the source tree.

Also refine the commit message rambling to be clearer about the why of
various requirements, such as the 50 (72) character header limit (GitHub
breaks them at 72). Give hints as to when one should better split up
their changes and when to use the commit message body. Explain what we
use a sign-off for and reiterate that it is not required.

As was recently the case, also mention what happens when there is a
larger batch of pull requests open and why we don't merge them
immediately one-by-one.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>